Move int pow/log into maths
