+++ pkg/proofs/ladder.go
@@ -1,65 +1,7 @@
 package proofs
 
-// ===========================================================================================
-// =====================================Log2Floor Lemmas======================================
-/*@
-// Function: Computes Log2Floor(n)
-ghost
-ensures r >= 0
-ensures n >= 1 ==> IntPow2(r) <= n && IntPow2(r+1) > n && r < n
-ensures n == 1 ==> r == 0
-decreases n
-pure
-func Log2Floor_pure(n uint64) (r uint64) {
-    return n < 2 ? 0 : 1 + Log2Floor_pure(n / 2)
-}
+import "github.com/felixlinker/keytrans-verification/pkg/utils"
 
-ghost
-requires a > 0
-requires b > 0
-requires a <= b
-ensures Log2Floor_pure(a) <= Log2Floor_pure(b)
-ensures a <= b/2 ==> Log2Floor_pure(a) < Log2Floor_pure(b)
-decreases b
-pure
-func Log2FloorMonotonic(a uint64, b uint64) uint64 {
-	return a == b ? 0 :
-		(b == 1 ? 0 :
-			(a == 1 ? Log2FloorMonotonic(1, b/2) :
-				Log2FloorMonotonic(a/2,b/2)))
-}
-
-@*/
-// ==================================================================================
-// =============================IntPow2 Lemmas ======================================
-
-/*@
-// Function: Computes 2^(exp)
-ghost
-requires exp >= 0
-ensures r > 0
-decreases exp
-pure
-func IntPow2(exp uint64) (r uint64) {
-  return exp == 0 ? 1 : 2 * IntPow2(exp - 1)
-}
-
-// Lemma: Weaker version of IntPow2IncLemma
-ghost
-requires a >= 0
-requires b >= 0
-requires a <= b
-ensures IntPow2(a) <= IntPow2(b)
-ensures a < b ==> IntPow2(a) < IntPow2(b)
-decreases b - a
-pure
-func IntPow2Monotonic(a uint64, b uint64) uint64 {
-	return a == b ? 0 : IntPow2Monotonic(a, b - 1)
-}
-
-@*/
-// ============================================================================
-// ============================================================================
 /*@
 ghost
 requires 0 <= t1 && 0 <= t2
@@ -90,8 +32,8 @@ ensures t1 != t2 ==> min(t1, t2) < r && r <= max(t1, t2)
 decreases
 pure func tStar_pure(t1 uint64, t2 uint64) (r uint64) {
 	return t1 == t2 ? 1 :
-		t1 < t2 ? let i_low := Log2Floor_pure(t1) in tStarRec_pure(t1, t2, IntPow2(i_low), IntPow2(i_low + 1)) :
-			let i_low := Log2Floor_pure(t2) in tStarRec_pure(t2, t1, IntPow2(i_low), IntPow2(i_low + 1))
+		t1 < t2 ? let i_low := utils.Log2Floor_pure(t1) in tStarRec_pure(t1, t2, utils.PowOf2_pure(i_low), utils.PowOf2_pure(i_low + 1)) :
+			let i_low := utils.Log2Floor_pure(t2) in tStarRec_pure(t2, t1, utils.PowOf2_pure(i_low), utils.PowOf2_pure(i_low + 1))
 }
 
 ghost
@@ -122,41 +64,10 @@ pure func tStarRec_pure_general(t1 uint64, t2 uint64, x_in uint64, x_out uint64)
 }
 @*/
 
-// @ requires base > 0
-// @ ensures r >= 0
-// @ ensures IntPow2(r) <= base
-// @ ensures base < IntPow2(r+1)
-// @ ensures r == Log2Floor_pure(base)
-// @ decreases base
-func Log2Floor(base uint64) (r uint64) {
-	if base <= 1 {
-		//@ assert Log2Floor_pure(base) == 0
-		return 0
-	} else {
-		//@ assert base > 1
-		//@ assert Log2Floor_pure(base) == 1 + Log2Floor_pure(base / 2)
-		return 1 + Log2Floor(base/2)
-	}
-}
-
-// @ requires exp >= 0
-// @ ensures r == IntPow2(exp)
-// @ decreases
-func PowOf2(exp uint64) (r uint64) {
-	var i uint64
-	r = 1
-	//@ invariant i >= 0 && i<= exp
-	//@ invariant r == IntPow2(i)
-	//@ decreases exp - i
-	for i = 0; i < exp; i += 1 {
-		r = r * 2
-	}
-	return r
-}
-
-// TStar returns a value r such that t1 < t2 ==> t1 < r <= t2
+// TStar returns a value r such that t1 < r <= t2
 // @ requires 0 <= t1 && 0 <= t2
-// @ ensures t_star == TStar_pure(t1,t2)
+// @ ensures t1 != t2 ==> 1 <= t_star
+// @ ensures t_star == TStar_pure(t1, t2)
 func TStar(t1 uint64, t2 uint64) (t_star uint64) {
 	return tStar(t1+1, t2+1) - 1
 }
@@ -167,11 +78,11 @@ func tStar(t1 uint64, t2 uint64) (t_star uint64) {
 	if t1 == t2 {
 		return 1
 	} else if t1 < t2 {
-		i_low := Log2Floor(t1)
-		return tStarRec(t1, t2, PowOf2(i_low), PowOf2(i_low+1))
+		i_low := utils.Log2Floor(t1)
+		return tStarRec(t1, t2, utils.PowOf2(i_low), utils.PowOf2(i_low+1))
 	} else {
-		i_low := Log2Floor(t2)
-		return tStarRec(t2, t1, PowOf2(i_low), PowOf2(i_low+1))
+		i_low := utils.Log2Floor(t2)
+		return tStarRec(t2, t1, utils.PowOf2(i_low), utils.PowOf2(i_low+1))
 	}
 }
 
@@ -224,7 +135,7 @@ func fullBinaryLadderSteps(target uint64 /*@, ghost t2 uint64@*/) (r []uint64 /*
 
 	/*@
 	ghost if t2 < target {
-		assert let _ := Log2FloorMonotonic(t2, target) in Log2Floor_pure(t2) <= Log2Floor_pure(target)
+		assert let _ := utils.Log2FloorMonotonic(t2, target) in utils.Log2Floor_pure(t2) <= utils.Log2Floor_pure(target)
 	}
 	@*/
 
@@ -234,8 +145,8 @@ func fullBinaryLadderSteps(target uint64 /*@, ghost t2 uint64@*/) (r []uint64 /*
 	// @ invariant acc(r)
 	// @ invariant len(r) == int(k)
 	// @ invariant 0 <= k && k <= target
-	// @ invariant i == IntPow2(k) && k == Log2Floor_pure(i)
-	// @ invariant 1 < k ==> k - 1 <= Log2Floor_pure(target)
+	// @ invariant i == utils.PowOf2_pure(k) && k == utils.Log2Floor_pure(i)
+	// @ invariant 1 < k ==> k - 1 <= utils.Log2Floor_pure(target)
 	// @ invariant forall j int :: 0 <= j && j < len(r) ==> 0 < r[j]
 	// @ invariant len(r) > 0 ==> r[k-1] == i / 2
 	// @ invariant 0 <= jump_idx
@@ -249,8 +160,8 @@ func fullBinaryLadderSteps(target uint64 /*@, ghost t2 uint64@*/) (r []uint64 /*
 
 		/*@
 		ghost if jump_idx == 0 && t2 < i {
-			assert let _ := Log2FloorMonotonic(i / 2, t2) in Log2Floor_pure(i / 2) <= Log2Floor_pure(t2)
-			assert let _ := Log2FloorMonotonic(t2, i) in Log2Floor_pure(t2) <= Log2Floor_pure(i)
+			assert let _ := utils.Log2FloorMonotonic(i / 2, t2) in utils.Log2Floor_pure(i / 2) <= utils.Log2Floor_pure(t2)
+			assert let _ := utils.Log2FloorMonotonic(t2, i) in utils.Log2Floor_pure(t2) <= utils.Log2Floor_pure(i)
 			jump_idx = len(r) - 1
 		}
 		@*/
@@ -268,8 +179,8 @@ func fullBinaryLadderSteps(target uint64 /*@, ghost t2 uint64@*/) (r []uint64 /*
 	x_out := i
 	r = append( /*@ perm(1/2), @*/ r, x_out) // this will be the first proof of non-inclusion
 
-	// @ assert let _ := Log2FloorMonotonic(target, i) in Log2Floor_pure(target) < k + 1
-	// @ assert let _ := Log2FloorMonotonic(k + 1, x_out) in k <= Log2Floor_pure(x_out)
+	// @ assert let _ := utils.Log2FloorMonotonic(target, i) in utils.Log2Floor_pure(target) < k + 1
+	// @ assert let _ := utils.Log2FloorMonotonic(k + 1, x_out) in k <= utils.Log2Floor_pure(x_out)
 
 	// Initialize the variables so that the recursive invariant on
 	// BinarySearchStep is established right-away when we already found t*.
@@ -291,7 +202,7 @@ func fullBinaryLadderSteps(target uint64 /*@, ghost t2 uint64@*/) (r []uint64 /*
 	// In the case t2 < target && jump_idx == 0, we must hint to gobra with
 	// assertion below that acc_x_in is indeed the argument it expects for the
 	// tStar_pure(t2, target) call in the function's post-condition.
-	// @ assert t2 < target && jump_idx == 0 ==> let _ := Log2FloorMonotonic(x_in, t2) in x_in == IntPow2(Log2Floor_pure(t2))
+	// @ assert t2 < target && jump_idx == 0 ==> let _ := utils.Log2FloorMonotonic(x_in, t2) in x_in == utils.PowOf2_pure(utils.Log2Floor_pure(t2))
 
 	return res /*@, search_idx @*/
 }
